move backend, nim_backend under utils
