+++ integrations/nvidia/src/haystack_integrations/components/embedders/nvidia/text_embedder.py
@@ -4,9 +4,8 @@ from urllib.parse import urlparse
 
 from haystack import component, default_from_dict, default_to_dict
 from haystack.utils import Secret, deserialize_secrets_inplace
+from haystack_integrations.util.nvidia import EmbedderBackend, NimBackend
 
-from ._nim_backend import NimBackend
-from .backend import EmbedderBackend
 from .truncate import EmbeddingTruncateMode
