+++ integrations/nvidia/src/haystack_integrations/util/nvidia/__init__.py
@@ -1,0 +1,4 @@
+from .backend import EmbedderBackend, GeneratorBackend, Model
+from .nim_backend import NimBackend
+
+__all__ = ["NimBackend", "EmbedderBackend", "GeneratorBackend", "Model"]
